Adding booklet files and images
This commit establishes a multi view booklet reading from a json file on the file system ( which will act as the default ). I want to add another update that will  read from a json file stored raw on github. I want to remove the navigation bar in this view. I need to add links to the patterns, and I want to tweak the images and spacing eventually.